feat: implement ExpenseTemplate component and integrate it into TransactionManager
